uploader/requirements.txt
anyio==4.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.5
mcrcon==0.7.0
oauthlib==3.2.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-telegram-bot==20.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==68.2.2
sniffio==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
